generalized some command-line arguments for r2f2u

+++ indexing/index.c
@@ -126,7 +126,7 @@ static void _index_recursive_write(FILE * output, ShardNode * node,
     if (depth > 0) {
         data[depth - 1] = node->nodeCharacter;
     }
-    unsigned int i;
+    long long i;
     if (node->depthRemaining == 0) {
         // just write the data, straightup
         for (i = 0; i < node->nodeDataSize; i++) {
